feat: Move the alpha portion of MSAA blend modes to the blend unit (#9576) 4f9625183f Move the alpha portion of MSAA blend modes to the blend unit
The advanced blend modes can be split algebraically into two distinct
operations: color and alpha.

The color portion has all the custom logic that needs to run in the
shader.

The alpha portion is run-of-the-mill source-over blending.

For the MSAA backend, move the alpha portion of the operation into the
hardware blend coefficients instead of doing it in the shader. This
gives smoother antialiasing because the blend unit does the blend at
each sample separately, whereas the shader runs once per fragment
against a single (resolved) destination color.