feat: First draft of a CommandQueue (#9620) 18dc7c390b * feat: First draft of a RiveCommandBuffer
RiveCommandBuffer hopes to eventually become the basic high-level Rive
API. It's an asynchronous command queue that hides the data behind
handles and executes on a separate thread.

This initial example is just something to get started. We will evolve
the API as we feel out what is needed in the various backends.

* More testing

* AutoLockAndNotify

* CommandQueue

* Fix linux build