fix(gl): Drop EXT_shader_pixel_local_storage (#10556) 8df4ac1a56 This extension has been an ever-flowing fountain of bugs for us, especially on PowerVR. Furthermore, it doesn't even work on Pixel 10 (PowerVR D-Series DXT-48-1536).
Let's de-fragment our Android GL renderer and consolidate everything on
MSAA (except for old PowerVR devices where MSAA doesn't work). MSAA is
hopefully a more tried-and-true, less bug-prone rendering mode in the
Android ecosystem anyway, and it's nice to be more consistent across
devices.

The perf story looks positive. On PowerVR, MSAA seems to be generally
faster. On ARM it's always been a mixed bag depending on content.

Either way, this is a nice approach to switch GL to maintenance mode so
we can go all in on Vulkan.

NOTE: We are still keeping EXT_shader_pixel_local_storage as a fallback
for the feather atlas on devices that don't support float rendering.
There are real GPUs where this is our only way to get high quality
feathers.